langgraph-server/requirements.txt
# LangGraph Server 依赖
langgraph>=0.2.0
langgraph-cli>=0.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# 数据库支持
psycopg2-binary>=2.9.0

# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# 工具和实用程序
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
